backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-jose==3.3.0
python-dotenv==1.0.0
boto3==1.26.165
requests==2.31.0
python-multipart==0.0.9
gunicorn==21.2.0
